web/requirements.txt
pip
apscheduler
redis
celery

psycopg2-binary
python-dotenv
requests

django
gunicorn
whitenoise

PyJWT

django-taggit
django-redis
django-jalali-date
django-cors-headers
django-admin-sortable2
django-import-export
djangorestframework
djangorestframework-jsonapi
djangorestframework-simplejwt
drf_spectacular
sentry-sdk
sentry-sdk[django]
sentry-sdk[celery]

debugpy
faker

ipython

singleton_package
telegram-bale-bot==4.18.4